+++ packages/common/ProForm/src/types/formCustomizerTypes.ts
@@ -89,6 +89,7 @@ export interface GroupSchema {
 	label?: string;
 	children: ProxyedSchema[];
 	grid?: GridStyle;
+	show?: boolean;
 }
 
 export interface ListSchema {
@@ -98,6 +99,7 @@ export interface ListSchema {
 	children: ProxyedSchema[];
 	grid?: GridStyle;
 	runtime?: Runtime;
+	show?: boolean;
 }
 
 export type Schema = ItemSchema | GroupSchema | ListSchema;