+++ packages/common/ProForm/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @harbor-design/proform
 
+## 1.1.29
+
+### Patch Changes
+
+- add show property to List and Group
+
 ## 1.1.28
 
 ### Patch Changes